requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
python-dotenv==1.0.0
twilio==8.10.3
openai==1.57.4
pydantic==2.5.0
python-multipart==0.0.6
aiohttp
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
psutil==5.9.6
pytest-cov
